SDESK-5048 Restructure publish tests.
