+++ tests/publish/fixtures/package.json
