+++ tests/publish/formatters/email_formatter_test.py
